Move the job provisioner tests into their own file
The six `TestExtractCloudInit` cases cover `JobProvisioner`, not the ssh provisioner whose file they
sat in.
